web/requirements.txt
Cython==0.22
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==1.0.8
WTForms==2.0.2
Werkzeug==0.10.4
awscli==1.9.17
boto==2.38.0
botocore==1.3.17
docutils==0.12
dominate==2.1.12
gunicorn==19.4.5
numpy==1.9.1
pyasn1==0.1.9
python-dateutil==2.4.2
rsa==3.2.3
scapy==2.3.1
virtualenv==13.1.0
visvis==1.9.1
web.py==0.37
